+++ doc/next/6-stdlib/99-minor/log/slog/62005.md
@@ -1,1 +1,1 @@
-DiscardHandler allows for the creation of a logger which discards all output with fewer resources consumed.
+DiscardHandler is a handler that is never enabled and always discards its output.
