--- src/log/slog/discard.go
@@ -1,14 +1,0 @@
-package slog
-
-import "context"
-
-// DiscardHandler discards all log output.
-// DiscardHandler.Enabled returns false for all Levels.
-var DiscardHandler Handler = discardHandler{}
-
-type discardHandler struct{}
-
-func (dh discardHandler) Enabled(context.Context, Level) bool  { return false }
-func (dh discardHandler) Handle(context.Context, Record) error { return nil }
-func (dh discardHandler) WithAttrs(attrs []Attr) Handler       { return dh }
-func (dh discardHandler) WithGroup(name string) Handler        { return dh }
